Inject HTTParty as default http client
HTTParty is now set as a named parameter on the HttpRequest::Get class
and set as the default, allowing easier switching out of a different
client in the future. It also helps to isolate the internals of the
wrapper class so it knows less about the internals of the client.