Add search spec helper and test data
New helper that takes a json file of recipe data and creates records for
testing. Some updates have been made to the factories although they are
not used in the spec helper – rather, the records are created directly.

The (unused in this branch) StubRequestHelper has been moved to be in
the conventional location for the spec helpers.