Find recipes with given ingredients
SearchController#search can now take a query and find recipes with a
list of ingredients. If a single ingredient is provided, it will find
recipes with that ingredient in the title, then the ingredients, then
the introduction, in that priority order.

If multiple ingredients are provided in the query, it will find recipes
containing all of those ingredients.